requirements.txt
# Chromosome + Cell Mini-Pipeline
# Python 3.10+ recommended

# Hi-C analysis
cooltools>=0.5.0
cooler>=0.9.0
bioframe>=0.4.0

# Microscopy / image analysis
stardist>=0.8.0
scikit-image>=0.22.0
scipy>=1.11.0
imageio>=2.33.0
tifffile>=2024.1.0

# Data & viz
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
matplotlib>=3.8.0

# Dashboard
streamlit>=1.40.0

# Config
pyyaml>=6.0
